+++ CHANGES.rst
@@ -6,6 +6,7 @@
   :meth:`pyudev.Monitor.enable_receiving()` now
 - :class:`pyudev.MonitorObserver` calls
   :meth:`pyudev.Monitor.enable_receiving()` now when started.
+- #40: Added user guide to the documentation
 
 
 0.14 (Feb 10, 2012)